Extract common http helpers to httpserver package
