+++ internal/httpserver/header.go
@@ -1,4 +1,4 @@
-package ingest
+package httpserver
 
 import (
 	"fmt"
@@ -8,15 +8,21 @@ import (
 )
 
 const (
-	mediaTypeJson = "application/json"
-	mediaTypeAny  = "*/*"
+	MediaTypeNDJson = "application/x-ndjson"
+	MediaTypeJson   = "application/json"
+	MediaTypeAny    = "*/*"
 )
 
-func enableCors(w http.ResponseWriter) {
+// EnableCors sets Access-Control-Allow-Origin to allow any origin.
+func EnableCors(w http.ResponseWriter) {
 	w.Header().Set("Access-Control-Allow-Origin", "*")
 }
 
-func acceptsAnyOf(w http.ResponseWriter, r *http.Request, strict bool, mts ...string) (string, bool) {
+// AcceptsMediaType checks whether the request Accept header matches one of the
+// given media types and returns the first match. When strict is false and
+// Accept is empty, it accepts. On mismatch or invalid Accept, it writes
+// StatusNotAcceptable and returns false.
+func AcceptsMediaType(w http.ResponseWriter, r *http.Request, strict bool, mts ...string) (string, bool) {
 	values := r.Header.Values("Accept")
 	if len(values) == 0 {
 		if !strict || len(mts) == 0 {
